+++ tests/testthat/test_as.regions.R
@@ -11,8 +11,6 @@ test_that(
     sc <- subset("GERMAPARLMINI", speaker == "Angela Dorothea Merkel")
     r2 <- as(sc, "regions")
     
-    r2@data_dir <- fs::path()
-    r1@data_dir <- fs::path()
     expect_identical(r1, r2)
   }
 )